+++ R/HFDutils.R
@@ -7,7 +7,6 @@
 # 4) getHFDitemavail()
 # 5) HFCparse()
 # 6) getHFCcountries()
-# 7) extract_HFD_items()
 
 ############################################################################
 # 1) HFDparse()
@@ -22,30 +21,30 @@
 #' @return DF same data.frame, modified so that columns are of a useful class. If there were open age categories, such as \code{"-"} or \code{"+"}, this information is stored in a new dummy column called \code{OpenInterval}.
 #' 
 #' @details This parse routine is based on the subjective opinions of the author...
-#' 
+#' @importFrom dplyr mutate relocate case_when last_col
+#' @importFrom readr parse_number
+#' @importFrom rlang .data
 #' @export
 #' 
 HFDparse <- function(DF){
 	if (any(c("Age","Cohort","ARDY") %in% colnames(DF))){
 		# assuming that if there are two such columns that the open age, etc, rows will always agree.    
-		DF$OpenInterval <- FALSE
+
 		if ("Age" %in% colnames(DF)){
-			Pluses     <- grepl(pattern = "\\+", DF$Age )
-			Minuses    <- grepl(pattern = "\\-", DF$Age )
-			DF$Age     <- age2int(DF$Age)    
-			DF$OpenInterval <- DF$OpenInterval | Pluses | Minuses
+			DF <- DF |> 
+			  mutate(Age = parse_number(.data$Age),
+			         OpenInterval = .data$Age %in% range(.data$Age))
 		}
 		if ("ARDY" %in% colnames(DF)){
-			Pluses     <- grepl(pattern = "\\+", DF$ARDY )
-			Minuses    <- grepl(pattern = "\\-", DF$ARDY )
-			DF$ARDY     <- age2int(DF$ARDY)    
-			DF$OpenInterval <- DF$OpenInterval | Pluses | Minuses
+		  DF <- DF |> 
+		    mutate(ARDY = parse_number(.data$ARDY),
+		           OpenInterval = .data$ARDY %in% range(.data$ARDY))
 		}
 		if ("Cohort" %in% colnames(DF)){
-			Pluses     <- grepl(pattern = "\\+", DF$Cohort )
-			Minuses    <- grepl(pattern = "\\-", DF$Cohort )
-			DF$Cohort  <- age2int(DF$Cohort)   
-			DF$OpenInterval <- DF$OpenInterval | Pluses | Minuses
+			DF <- DF |>
+			  mutate(OpenInterval = grepl(pattern = "\\+", .data$Cohort) |  grepl(pattern = "\\-", .data$Cohort) ,
+			         Cohort = parse_number(.data$Cohort)) |>
+			  relocate(.data$OpenInterval, .after = last_col())
 		}
 	}
 	DF
@@ -57,25 +56,52 @@ HFDparse <- function(DF){
 
 #' @title internal function for grabbing the HFD country short codes. 
 #'
-#' @description This function is called by \code{readHFDweb()} and is separated here for modularity. There is likely a simpler way of coding this functionality. The vector of short codes returned only includes the fully incorporated HFD countries, not provisional countries.
+#' @description This function is called by \code{readHFDweb()} and is separated here for modularity. We include both main and provisional countries in the grab.
 #' 
-#' @return a vector of HFD country short codes.
+#' @return a `tibble` with three columns `Country`, `link` and `CNTRY` (the country short code)
 #' 
-#' @importFrom XML readHTMLTable
-#' @importFrom httr GET
+#' @importFrom rvest read_html html_element html_elements html_attr html_text2
+#' @importFrom dplyr tibble mutate
+#' @importFrom rlang .data
 #' 
 #' @export
 #' 
 getHFDcountries <- function(){
-	# the ugliest code I've ever written. There must be a better way..
-  hfd_url <- "https://www.humanfertility.org/cgi-bin/zipfiles.php"
-  some_html <- httr::GET(hfd_url)
-  X <- XML::readHTMLTable(XML::htmlParse(some_html),
-                          header = TRUE, 
-                          colClasses = c("character", "character"), 
-                          which = 2, 
-                          stringsAsFactors = FALSE)
-	gsub("\\s*\\([^\\)]+\\)", "", X[,2])
+
+  # xpath to the main country list
+  xpath <- "/html/body/div[1]/div/div[3]/div[3]"
+  
+  html <- read_html("https://www.humanfertility.org")
+  links <-
+    html |>
+    html_element(xpath=xpath) |>
+    html_elements("a") |>
+    html_attr("href")
+  
+  cntry_names <-
+    html |> 
+    html_element(xpath=xpath) |> 
+    html_elements("a") |> 
+    html_text2()
+  
+  # preliminary releases are in a separate list
+  xpath_prelim <- "/html/body/div[1]/div/div[4]/div[3]"
+  links2 <-
+    html |> 
+    html_element(xpath=xpath_prelim) |> 
+    html_elements("a") |> 
+    html_attr("href")
+  
+  cntry_names2 <-
+    html |> 
+    html_element(xpath=xpath_prelim) |> 
+    html_elements("a") |>
+    html_text2()
+  
+  # compose table and extract country code from links:
+  tibble(Country= c(cntry_names,cntry_names2), link = c(links, links2)) |> 
+    mutate(CNTRY =  sub(".*=", "", .data$link))
+  
 }
 
 
@@ -91,59 +117,140 @@ getHFDcountries <- function(){
 #' 
 #' @return character string of eight integers representing the date as \code{"yyyymmdd"}.
 #' 
-#' @importFrom httr GET content
+#' @importFrom rvest read_html html_elements html_text2
+#' @importFrom lubridate dmy year month day
+#' @importFrom stringr str_pad
 #' 
 #' @export
 #' 
 getHFDdate <- function(CNTRY){
-  CountryURL      <- paste0("http://www.humanfertility.org/cgi-bin/country.php?",
-                            "country=", CNTRY)
-  #get date string by finding the permalink listed at the bottom of page.
-  html_country    <- httr::GET(CountryURL)
-  htmlstr <- httr::content(html_country, as="text")
-  LastUpdate <- regmatches(htmlstr, regexpr("(?<=update=)20[0-9]{6}", htmlstr,
-                                            perl=T))
+  CountryURL <- paste0("https://www.humanfertility.org/Country/Country?cntr=", CNTRY)
+  html       <- read_html(CountryURL)
+  xpath      <- "/html/body/div[1]/div/div[3]/div[1]/div[1]/div[2]/span"
+
+  LastUpdate <- 
+    html |>  
+    html_elements(xpath = xpath) |> 
+    html_text2() |>
+    sub(pattern = ".*: ", replacement = "", ) |>
+    dmy()
+  
   if(length(LastUpdate)==0){
     stop("I can't find the date of the latest update to the data for this
           country. The Human Fertility Database website may have changed")
   }
+  date_out <- paste0(year(LastUpdate),
+                     str_pad(month(LastUpdate),
+                             width = 2,
+                             side="left",
+                             pad="0"),
+                     str_pad(day(LastUpdate),
+                             width = 2,
+                             side="left",
+                             pad="0"))
+  
+  
   # this isn't a date string, just 8 digits squashed together yyyymmdd
-  LastUpdate
+  date_out
 }
+
 
 
 ############################################################################
 # 4) getHFDitemavail()
 ############################################################################
 
-#' @title internal function for grabbing the available data item names for a given country.
+#' @title List the available data item names for a given HFD country.
 #' 
 #' @description called by \code{readHFDweb()}. This assumes that \code{CNTRY} is actually available in the HFD. 
 #' 
 #' @param CNTRY HFD country short code.
 #' 
-#' @return a vector of item names. These are the file base names, and only need the extension \code{.txt} added in order to get the file name.
+#' @return a tibble of all available data files for the selected country. There are several useful indentifiers that can help determine the appropriate file, including the `measure` and `subtype` as detected from the html table properties, and `lexis` and `parity` as detected either from the file names or the table properties.
 #' 
-#' @importFrom httr GET content
+#' @importFrom janitor clean_names
+#' @importFrom tidyr pivot_longer
+#' @importFrom dplyr rename filter bind_rows bind_cols mutate case_when if_else select
+#' @importFrom rvest read_html html_table html_elements html_text2 html_attr
+#' @importFrom stringr str_split
+#' @importFrom rlang .data
 #' 
 #' @export
 #' 
 getHFDitemavail <- function(CNTRY){
-  # It seems this function will only worked if you are logged in
-	CountryURL      <- paste0("https://www.humanfertility.org/cgi-bin/",
-	                          "country.php?country=",CNTRY)
+ 
+  CountryURL <- paste0("https://www.humanfertility.org/Country/Country?cntr=", CNTRY)
+  
+  tidy_chunk <- function(X){
+    X |>
+      clean_names() |>
+      rename("measure" = .data$x) |> 
+      pivot_longer(-.data$measure,names_to = "subtype",values_to = "years") |> 
+      filter(.data$measure != "")
+  }
+  
+  html <- read_html(CountryURL)
+  
+  cntry_tables<-
+    html |>
+    html_table() |>
+    lapply(FUN = tidy_chunk) |>
+    bind_rows() |>
+    filter(years != "-")
+  
+  years <-
+    html |>
+    html_elements("table") |>
+    html_elements("tr")|>
+    html_elements("a") |>
+    html_text2()
+  
+  links <-
+    html |>
+    html_elements("table") |>
+    html_elements("tr")|>
+    html_elements("a") |>
+    html_attr("href")
+  
+  # years2 was a check to ensure join was row-matched properly
+  linksyears <- tibble(link = links, years2 = years)
+  
+  # just need to infer some identifier columns based mostly on file names,
+  # and similar inferences
+  item_table <- 
+  cntry_tables |>
+    bind_cols(linksyears) |>
+    filter(grepl(.data$link,pattern="*.txt")) |>
+    mutate(lexis = case_when(grepl(.data$link,pattern = "TR") ~ "triangle",
+                             grepl(.data$link,pattern = "RR") ~ "age-period",
+                             grepl(.data$link,pattern = "VH") ~ "age-cohort",
+                             grepl(.data$link,pattern = "VV") ~ "period-cohort",
+                             grepl(.data$link,pattern = "pa") ~ "period",
+                             grepl(.data$link,pattern = "pft") ~ "period",
+                             grepl(.data$link,pattern = "patfr") ~ "period",
+                             grepl(.data$link,pattern = "pmab") ~ "period",
+                             grepl(.data$link,pattern = "pmabc") ~ "period",
+                             grepl(.data$link,pattern = "cft") ~ "cohort",
+                             grepl(.data$link,pattern = paste0(CNTRY,"mi")) ~ "period",
+                             grepl(.data$link,pattern =  paste0(CNTRY,"mic")) ~ "period",
+                             grepl(.data$link,pattern =  paste0(CNTRY,"births")) ~ "mixed",
+                             grepl(.data$link,pattern =  paste0(CNTRY,"monthly")) ~ "period",
+                             grepl(.data$link,pattern =  paste0(CNTRY,"parity")) ~ "mixed"), 
+    subtype = gsub('[0-9]+', '', .data$subtype),
+           subtype = sub("_$", "", .data$subtype),
+           subtype = if_else(.data$subtype == "years", "Input data", .data$subtype)) |> 
+  select(-.data$years2) |>
+  mutate(item = .data$link |>
+           str_split(pattern = "\\\\") |>
+           lapply(rev) |>
+           lapply('[',1) |>
+           unlist() |>
+           gsub(pattern = ".txt", replacement = "") |>
+           gsub(pattern = CNTRY, replacement = ""),
+         link = gsub(.data$link, pattern = "\\\\", replacement = "/"),
+         link = paste0("https://www.humanfertility.org", .data$link))
 
-	# vector of names of tabs on each HFD page
-	tabs <- c("si", "asfr", "ft")
-
-	item_list <- lapply(tabs, function(tab){
-	  # iterate through each tab and construct URL
-	  url <- paste0(CountryURL, "&tab=", tab)
-	  tab_html  <- httr::content(httr::GET(url))
-	  return(extract_HFD_items(tab_html, CNTRY))
-	})
-	
-	return(unlist(item_list))
+  item_table
 }
 
 
@@ -223,36 +330,6 @@ getHFCcountries <- function(names = FALSE){
 }
 
 
-
-############################################################################
-# 7) extract_HFD_items()
-############################################################################
-#' @title Extract names of HFD items from HTML
-#'
-#' @description Extract names of Human fertility database datasets from an
-#' HTML page. Links are extracted from html, and for each link, any text falling
-#' between the provided country code and a .txt extension is returned
-#'
-#' @param html_content An object of class \code{xml_document} taken from a
-#' country page of the HFD website, typically the result of
-#' \code{httr::content(httr::GET(url))}
-#'
-#' @param CNTRY A string given one of the country codes used by HFD. See
-#' \code{getHFDcountries} for a function which lists available countries.
-#'
-#' @return A vector of characters, each of which is an item available for
-#' download
-#'
-#' @importFrom XML getHTMLLinks htmlParse
-#'
-extract_HFD_items <- function(html_content,CNTRY) {
-  parsed_html <- XML::htmlParse(html_content)
-  all_links <- XML::getHTMLLinks(parsed_html)
-  item_regex <- regexpr(paste0("(?<=",CNTRY,")[A-Za-z0-9]+(?=\\.txt)"), 
-          all_links, perl=T)
-  item_names <- regmatches(all_links, item_regex)
-  return(item_names)
-}