+++ R/readHFD.R
@@ -29,8 +29,8 @@
 #' @note original function submitted by Josh Goldstein, modified by Tim Riffe.
 #' 
 
-readHFD <- function(filepath, fixup = TRUE,...){
-    DF      <- read.table(file = filepath, header = TRUE, skip = 2, na.strings = ".", as.is = TRUE, ...)
+readHFD <- function(filepath, fixup = TRUE, ...){
+    DF      <- suppressWarnings(read.table(file = filepath, header = TRUE, skip = 2, na.strings = ".", as.is = TRUE, ...))
     if (fixup){
       DF      <- HFDparse(DF)
     }
@@ -44,7 +44,7 @@ readHFD <- function(filepath, fixup = TRUE,...){
 #'
 #' @title read an HFD data file directly from the web as an R data.frame
 #' 
-#' @description Read HFD data directly from the web. This function is useful for short reproducible examples, or to make code guaranteed to always use the most up to date version of a particular HFD data file. For working with the entire HFD for a comparative study, it may be more efficient to download the full HFD and read in the elements using \code{readHFD()}. This function returns data formatted in the same way as \code{readHFD()}, that is, with Age columns (and others) converted to integer, and with open age group identifiers stored in a new logical column called \code{OpenInterval}. This reduces user burden somewhat, and facilitates direct use of functionality such as \code{log2mat()}. It is faster to specify \code{CNTRY} and \code{item} as arguments than to make the function figure out what's available. For repeated calls to this function, you can pass your username and password in as variables without having to include these in you R script by using \code{userInput()}-- see example. The user also has the option of querying particular updates from the HFD revision history. If you wish to specify a particular update, you must know the date that a particular country was updated, in the format \code{"YYYYMMDD"}. These dates differ between countries, so keep a good record if you wish your work to be reproducible to that extent (as well as lightweight)!
+#' @description Read HFD data directly from the web. This function is useful for short reproducible examples, or to make code guaranteed to always use the most up to date version of a particular HFD data file. For working with the entire HFD for a comparative study, it may be more efficient to download the full HFD zip files and read in the elements using \code{readHFD()}. This function returns data formatted in the same way as \code{readHFD()}, that is, with Age columns (and others) converted to integer, and with open age group identifiers stored in a new logical column called \code{OpenInterval}. It is faster to specify \code{CNTRY} and \code{item} as arguments than to make the function figure out what's available. For repeated calls to this function, you can pass your username and password in as variables without having to include these in you R script by using \code{userInput()}-- see example. The user also has the option of querying particular updates from the HFD revision history. If you wish to specify a particular update, you must know the date that a particular country was updated, in the format \code{"YYYYMMDD"}. These dates differ between countries, so keep a good record if you wish your work to be reproducible to that extent (as well as lightweight)!
 #' 
 #' @param CNTRY character string of the HFD short code. Only one!
 #' @param item character string of the data product code, which is the base file name, but excluding the country code and file extension \code{.txt}. For instance, \code{"mabRR"} or \code{"tfrVHbo"}. If you're not sure, then leave it blank and a list will be shown. Only one item!
@@ -57,9 +57,12 @@ readHFD <- function(filepath, fixup = TRUE,...){
 #'
 #' @details You need to register for HFD to use this function: \url{https://www.humanfertility.org}. It is advised to pass in your credentials as named vectors rather than directly as character strings, so that they are not saved directly in your code. See examples. One option is to just save them in your Rprofile file.
 #' 
-#' @importFrom httr GET content
+#' @importFrom httr content status_code
+#' @importFrom rvest session html_form html_form_set session_submit session_jump_to  
 #' @importFrom utils select.list
-#' @importFrom utils read.table
+#' @importFrom dplyr filter pull
+#' @importFrom lubridate ymd is.Date
+
 #' 
 #' @export
 #' 
@@ -83,7 +86,13 @@ readHFD <- function(filepath, fixup = TRUE,...){
 #' ### # plus enter data in the console twice:
 #' ### DAT <- readHFDweb()
 #' 
-readHFDweb <- function(CNTRY = NULL, item = NULL, username = NULL, password = NULL, fixup = TRUE, Update = NULL){
+readHFDweb <- function(
+    CNTRY = NULL, 
+    item = NULL, 
+    username = NULL, 
+    password = NULL, 
+    fixup = TRUE, 
+    Update = NULL){
 	
 	# let user input name and password
 	if (is.null(username)){
@@ -114,51 +123,80 @@ readHFDweb <- function(CNTRY = NULL, item = NULL, username = NULL, password = NU
 			stop("CNTRY should be one of these:\n",paste(CNTRIES, collapse = ",\n"))
 		}
 	}
-	
-	# concatenate the login string
-	loginURL <- paste0("https://www.humanfertility.org/cgi-bin/logon.plx",
-	                   "?page=main.php&f=na&tab=na&LogonEMail=",
-			username, "&LogonPassword=", password, "&Logon=%20%20Login%20%20%20"
-	)
 
-	logged_in_page <- httr::GET(loginURL)
-	Continue <- grepl("welcome", logged_in_page)
-	if (!Continue) {
-	  stop(paste0("login didn't work. \nMaybe your username or password are off?",
-	              " \nYour request is contracepted!"))
-  }
-	# let user chose, or filter items as necessary: 
+  
+  # testing starts here
+  loginURL <- paste0("https://www.humanfertility.org/Account/Login")
+	# concatenate the login string
+
+	html <- session(loginURL)
 	
-	items <- getHFDitemavail(CNTRY)
-	if(is.null(item) || !(item %in% items)){
+	# olny one form on login page:
+	pgform    <- html_form(html)[[1]]  
+	
+	# # hack because rvest doesn't record where we were??
+	pgform$action <- loginURL
+	pgform$url    <- loginURL
+	
+	filled_form   <- html_form_set(pgform, 
+	                               Email = username, 
+	                               Password = password)
+
+	# TR: yay this now works!
+	html2 <- session_submit(html, filled_form)
+
+	Continue <- status_code(html2) == 200
+	if (!Continue) {
+	  stop(paste0("login didn't work. \nMaybe your username or password are off?
+If your username and password are from before 4 November 2022
+then you'll need to re-register for HFD, starting here:\n
+https://www.humanfertility.org/Account/UserAgreement"))
+  }
+	# let user choose, or filter items as necessary: 
+
+	itemavail <- getHFDitemavail(CNTRY)
+	items     <- itemavail$item
+	if(is.null(item) || !(item %in% items) | length(item) > 1){
 		if (interactive()){
-			item <- select.list(choices = items, multiple = FALSE, title = "Select item")
+		  
+		  cat(paste0("select an item nr\nIf you're not sure which one you want, select 0 and try running\ngetHFDitemavail('",eval(CNTRY),"') |> View()"))
+			.item <- select.list(choices = items, multiple = FALSE, title = "Select item")
 		} else {
 			stop("item should be one of these:\n",paste(item, collapse = ",\n"))
 		}
+	} else {
+	  .item = item
 	}
 	
-	# everything is in a folder, the name of which is the 8-digit version of the update date,
-  # user could specify a different number. Don't know how to query what dates are available, 
-  # though. by default we just get the most recent date.
-  if (is.null(Update)){
-		Update <- getHFDdate(CNTRY)
-	}	
+	# data_url <- itemavail |> filter(item == .item) |> pull(link)
+	
+	if (is.null(Update)){
+	  yyyymmdd <- getHFDdate(CNTRY)
+	} else {
+	  if (is.Date(ymd(yyyymmdd))){
+	    cat("Attempting to retrieve data from your requested Update date")
+	    yyyymmdd <- Update
+	  } else {
+	    stop("Update date appears misspecified. It needs to be a string in 'yyyymmdd' format\n The most recent valid update for ",CNTRY, " was '",eval(yyyymmdd),"'")
+	  }
+	}
+	# this could be pulled from getHFDitemavail() output as well,
+	# but we self-construct in order to allow custom dates
+	grab_url  <- paste0("https://www.humanfertility.org/File/GetDocument/Files/",CNTRY,
+	                    "/",yyyymmdd,"/",CNTRY,.item,".txt")
+	data_grab <- session_jump_to(html2, url = grab_url)
+	tmp <- tempfile()
+	
+	data_grab$response |> 
+	  content(encoding = "UTF-8") |> 
+	  cat(file=tmp, encoding = "UTF-8")
+	
+	DF <- readHFD(tmp, fixup = fixup)
+	
+	unlink(tmp)
+	closeAllConnections()
 
-	# url used to ask for data file. try to ignore 'tabs'
-	HFDurl <- paste0(
-			"https://www.humanfertility.org/cgi-bin/getfile.plx?f=",
-			CNTRY, "\\", Update, "\\", CNTRY, item, ".txt&c=", CNTRY)
-	
-	
-	Text <- httr::GET(HFDurl)
-	# parse raw data file to data.frame
-	DF <- try(read.table(text = httr::content(Text,encoding = "UTF-8"), header = TRUE, skip = 2, 
-	                     na.strings = ".", as.is = TRUE), silent = TRUE)
-	
-	if (fixup){
-		DF      <- HFDparse(DF)
-	}
+
 	return(invisible(DF))
 	
 }